requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
diff-match-patch==20230430
Django==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-import-export==4.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
fast-depends==2.4.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
lxml==5.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
